// src/components/types.ts
export type ColorVariant =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'danger'
  | 'warning'
  | 'info'
  | 'light'
  | 'dark'
  | 'neutral';

export type SizeVariant = 'xs' | 'sm' | 'md' | 'lg' | 'xl';

export type FlexAlign = 'flex-start' | 'center' | 'flex-end' | 'space-between';

export type DirectionType = 'row' | 'column';